[vxworks] [ppc] match TARGET_VXWORKS64 to TARGET_64BIT
Configuring gcc for --target=powerpc-wrs-vxworks7r2 sets things up for
a 64-bit compiler, just like powerpc64-wrs-vxworks7r2, except that
TARGET_VXWORKS64 is only defined as 1 for targets that match
*64-*-vxworks*.

With !TARGET_VXWORKS64, we get a 64-bit toolchain that defines
SIZE_TYPE, PTRDIFF_TYPE, and WCHAR_TYPE as 32-bit types, and that
breaks GCC passes that expect SIZE_TYPE and PTRDIFF_TYPE to be as wide
as pointers.

Arrange for TARGET_VXWORKS64 on ppc to match TARGET_64BIT, after using
it to select the default word size with driver self specs.


for  gcc/ChangeLog

	* config/rs6000/vxworks.h (SUBTARGET_DRIVER_SELF_SPECS):
	Redefine to select word size matching TARGET_VXWORKS64.
	(TARGET_VXWORKS64): Redefine in terms of TARGET_64BIT.