ifconv: Remove unused array predicated
While starting to improve if-conv, I noticed that predicated
was only being set once inside a loop and accessed right below.
This became this way due to r14-8869-g8636c538b68068 and before
it was needed since it was accessed via 2 loops but now it is only set
and then accessed in the next statement, there is no reason for it being
there.  So let's remove it and just use the value from it instead.

Bootstrapped and tested on x86_64-linux-gnu.

gcc/ChangeLog:

	* tree-if-conv.cc (combine_blocks): Remove predicated
	dynamic array.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>