fortran: Add the preliminary code of MOVE_ALLOC arguments
Add the preliminary code produced for the evaluation of the FROM and TO
arguments of the MOVE_ALLOC intrinsic before using their values.
Before this change, the preliminary code was ignored and dropped,
limiting the validity of the implementation of MOVE_ALLOC to simple
cases without preliminary code.

This change also adds the cleanup code of the same arguments.  It
doesn't make any difference on the testcase though.  Because of the
limited set of arguments that are allowed (variables or components
without subreference), it is possible that the cleanup code is actually
guaranteed to be empty.  At least adding the cleanup code makes the
array case consistent with the scalar case.

gcc/fortran/ChangeLog:

	* trans-intrinsic.cc (conv_intrinsic_move_alloc): Add pre and
	post code for the FROM and TO arguments.

gcc/testsuite/ChangeLog:

	* gfortran.dg/move_alloc_20.f03: New test.