ifconv: Small improvement to fold_build_cond_expr; lhs and rhs being the same.
This is a small compile time optimization, as match and simplify will generate
the same thing but with rhs and lhs being the same we can return early instead
of having to go through match and simplify. This might not show up that much
at this point but can/will show up after my patch for PR 119920 where we factor
out common code between the 2 sides of the if statement while in if-conv.

Bootstrapped and tested on x86_64-linux-gnu.

gcc/ChangeLog:

	* tree-if-conv.cc (fold_build_cond_expr): Return early if lhs and rhs
	are the same.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>