ada: Additional condition for Capacity discriminant on bounded container aggregates
This change test an additional condition as part of the criteria used
for deciding whether to generate a call to a container type's Length
function (for passing to the Empty function) when determining the
size of the object to allocate for a bounded container aggregate
with a "for of" iterator.

An update is also made to function Empty in Ada.Containers.Bounded_Hash_Maps,
adding a default to the formal Capacity, to make it consistent with other
bounded containers (and to make it conformant with the Ada RM).

gcc/ada/ChangeLog:

	* libgnat/a-cbhama.ads (Empty): Add missing default to Capacity formal.
	* libgnat/a-cbhama.adb (Empty): Add missing default to Capacity formal.
	* exp_aggr.adb (Build_Size_Expr): Test for presence of Capacity
	discriminant as additional criterion for generating the call to
	the Length function. Update comments.