RISC-V: Ignore non-types in builtin function hash.
If a user passes a string that doesn't represent a variable we still try
to compute a hash for its type.  Its tree does not represent a type but
just an exceptional, though.  This patch just ignores it, leaving the
error to the checking code later.

	PR target/113829

gcc/ChangeLog:

	* config/riscv/riscv-vector-builtins.cc (registered_function::overloaded_hash):
	Skip non-type arguments.

gcc/testsuite/ChangeLog:

	* gcc.target/riscv/rvv/base/pr113829.c: New test.